+++ src/main/java/svnserver/config/GitRepositoryConfig.java
@@ -12,13 +12,11 @@ import org.eclipse.jgit.lib.Repository;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.tmatesoft.svn.core.SVNException;
 import svnserver.context.LocalContext;
 import svnserver.repository.git.GitCreateMode;
 import svnserver.repository.git.GitLocation;
 import svnserver.repository.git.GitRepository;
 import svnserver.repository.git.push.GitPushEmbeddedConfig;
-import svnserver.repository.locks.PersistentLockFactory;
 
 import java.io.File;
 import java.io.IOException;
@@ -52,7 +50,7 @@ public final class GitRepositoryConfig {
   }
 
   @NotNull
-  public GitRepository create(@NotNull LocalContext context) throws IOException, SVNException {
+  public GitRepository create(@NotNull LocalContext context) throws IOException {
     return create(context, ConfigHelper.joinPath(context.getShared().getBasePath(), path));
   }
 
@@ -60,7 +58,7 @@ public final class GitRepositoryConfig {
   public GitRepository create(@NotNull LocalContext context, @NotNull File fullPath) throws IOException {
     context.add(GitLocation.class, new GitLocation(fullPath));
 
-    return new GitRepository(context, createRepository(context, fullPath), pusher.create(context), branch, renameDetection, new PersistentLockFactory(context));
+    return new GitRepository(context, createRepository(context, fullPath), pusher.create(context), branch, renameDetection);
   }
 
   @NotNull